Add room name to url
